src/requirements.txt
Django==3.2.8
djangorestframework>=3.12.2
whitenoise==5.3.0
django-environ==0.7.0
django-extensions==3.1.0
django-celery-results==2.2.0
django-split-settings==1.1.0

python-telegram-bot==12.8
seller-stats==0.3.1

scrapinghub==2.3.1
sentry-sdk==1.4.3

openpyxl==3.0.9
PyYAML==6.0
pandas==1.3.3
envparse==0.2.0
msgpack==1.0.2
json_log_formatter==0.4.0
python-dateutil==2.8.2

celery==5.1.2
flower==1.0.0
redis==3.5.3
boto3==1.18.58
psycopg2-binary==2.9.1
psutil==5.8.0


pytest-django==4.4.0
pytest-cov==3.0.0
pytest-env==0.6.2
mixer==7.1.2
coverage==6.0.1
codecov==2.1.12
freezegun==1.1.0
mongomock==3.23.0
Faker==9.2.0
requests_mock==1.9.3

typing==3.7.4.3
typing-extensions==3.10.0.2

jedi==0.18.0
flake8==3.9.2
flake8-bugbear==21.9.2
flake8-commas==2.0.0
flake8-isort==4.0.0
flake8-print==4.0.0
flake8-blind-except==0.2.0
flake8-builtins==1.5.3
flake8-comprehensions==3.6.1
flake8-debugger==4.0.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.0
flake8-todo==0.7
flake8-annotations-complexity==0.0.6
flake8-variables-names==0.0.4
flake8-broken-line==0.3.0
flake8-absolute-import==1.0
flake8-django==1.1.2
flake8-eradicate==1.1.0
flake8-mock==0.3
flake8-pie==0.15.0
flake8-printf-formatting==1.1.2
flake8-pytest==1.3
flake8-pytest-style==1.5.0